Create MutliFilter Search_Single table
Converting a string into a temporary table to compare against a value. For example, when searching for multiple cities, a comma separated string  will be assigned to @UserPref. This will be turned into a temporary table with one column. It will then be compared  to the city column for each listing.